Prework + Tutor
forgot to push before my tutoring session again, whoops! My (now somewhat redundant) work can be seen above the function in the script. I got the date display too. The for loop came with the help of my tutor, Phillip Loy, who got me to take advantage of the child relationships present in the HTML to take care of the local storage of the text as well as updating the time classes in one go.